+++ SystemeDeReservation/src/main/webapp/WEB-INF/views/listeVoiture.jsp
@@ -31,6 +31,11 @@ body {
 <body>
 <jsp:include page="header.jsp"></jsp:include>
 <jsp:include page="menuVoiture.jsp"></jsp:include>
+
+<c:if test="${!empty messages}">
+<span class="label label-warning"><h2>${messages}</h2></span>
+</c:if>
+
 <div class="container">
 	<div class="row">
         <div class="span12">
@@ -50,7 +55,7 @@ body {
 	                      			</h4>
 	                    			</td>
 	                    			<td>		
-		                   			<a href="${pageContext.request.contextPath}/voiture/edit/${voiture.id}" class="btn btn-primary icon  pull-right">Editer</a>                    		
+		                   			<a href="${pageContext.request.contextPath}/voiture/edit/${voiture.id}" class="btn btn-primary icon  pull-right">Modifier</a>                    		
 		                    		</td>
                    		 </tr>
                    		 <tr>
@@ -60,7 +65,7 @@ body {
 	                      			</h4>
 	                      			</td>
 	                      			<td>
-		                    		<a href="${pageContext.request.contextPath}/voiture/delete/${voiture.id}" class="btn btn-primary icon  pull-right">Supprimer</a>
+		                    		<a href="${pageContext.request.contextPath}/voiture/delete/${voiture.id}" class="btn btn-primary icon  pull-right"  onclick="return confirm('Voulez vous vraiment supprimer ?')">Supprimer</a>
 		                     		</td>
                     	 </tr>
                     </div>
